rename fetch.py -> sync.py
